feat(product): implement access verification by user level

+++ src/product/product.controller.ts
@@ -2,6 +2,8 @@ import { Controller, Get, Post, Body, Patch, Param, Delete } from '@nestjs/commo
 import { ProductService } from './product.service';
 import { CreateProductDto } from './dto/create-product.dto';
 import { UpdateProductDto } from './dto/update-product.dto';
+import { CurrentUser } from 'src/auth/decorators/current-user.decorator';
+import { User } from 'src/user/entities/user.entity';
 
 @Controller('product')
 export class ProductController {
@@ -28,7 +30,9 @@ export class ProductController {
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
-    return this.productService.remove(+id);
+  remove(@Param('id') id: string, @CurrentUser() user: User) {
+    if (user.accessLevel === 1) {
+      return this.productService.remove(+id);
+    }
   }
 }
